Arrumado o bug do u8 (dica: não usem long).

+++ JVM_Psicologia/src/ClassFile.h
@@ -17,7 +17,7 @@
 typedef unsigned char u1;
 typedef unsigned short u2;
 typedef unsigned int u4;
-typedef unsigned long u8;
+typedef unsigned long long u8;
 
 //Tabela dos constant_infos
 #define CONSTANT_Class				7
